fix: add explicit cast in get_stride to eliminate C4244 warning (#4759)
The get_stride() function returns uint32_t but all.weights.stride() returns
uint64_t, causing implicit conversion warnings when external consumers compile
against the public API headers (e.g., in NuGet test projects).

Added static_cast<uint32_t> to make the conversion explicit and intentional,
matching the pattern used in num_weights() on line 339.

This eliminates the remaining 8 C4244 warnings that appear in NuGet tests:
- 4 warnings in v141 (VS 2017) tests
- 4 warnings in v142 (VS 2019) tests

File: vowpalwabbit/core/include/vw/core/vw.h:341

+++ vowpalwabbit/core/include/vw/core/vw.h
@@ -338,7 +338,7 @@ inline void set_weight(VW::workspace& all, uint32_t index, uint32_t offset, floa
 
 inline uint32_t num_weights(VW::workspace& all) { return static_cast<uint32_t>(all.length()); }
 
-inline uint32_t get_stride(VW::workspace& all) { return all.weights.stride(); }
+inline uint32_t get_stride(VW::workspace& all) { return static_cast<uint32_t>(all.weights.stride()); }
 
 inline void init_features(primitive_feature_space& fs, size_t features_count)
 {
